backend/requirements.txt
fastapi
uvicorn
torch
transformers
numpy
nltk
pydantic
python-multipart
python-docx
pypdf
